+++ tests/models/test_certificate.py
@@ -1,8 +1,12 @@
 from __future__ import annotations
 
+from datetime import datetime
+from datetime import timezone
+from distutils.version import LooseVersion
 from unittest import mock
 from unittest.mock import PropertyMock
 
+import cryptography
 import pytest
 from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives import serialization
@@ -94,6 +98,14 @@ def certificate(certificate_asn1) -> Certificate:
     return Certificate.from_ans1(certificate_asn1)
 
 
+@pytest.mark.skipif(
+    LooseVersion(cryptography.__version__) < LooseVersion('36.0.0'),
+    reason=(
+        f'Cryptography {cryptography.__version__} < 36.0.0 generates '
+        f'slightly different CSR public bytes for certificate request '
+        f'than is expected here.'
+    ),
+)
 def test_create_certificate_signing_request(unencrypted_pem, certificate_signing_request_pem):
     pk = PrivateKey.from_pem(unencrypted_pem.content)
     csr = Certificate.create_certificate_signing_request(pk)
@@ -173,3 +185,59 @@ def test_is_development_certificate(is_development_cert, cert_common_name, certi
     patched_common_name = PropertyMock(return_value=cert_common_name)
     with mock.patch.object(Certificate, 'common_name', new_callable=patched_common_name):
         assert certificate.is_development_certificate is is_development_cert
+
+
+def test_subject(certificate):
+    expected_subject = {
+        'C': 'US',
+        'CN': 'iPhone Developer: Created via API (6HQ443353U)',
+        'O': 'NEVERCODE LTD',
+        'OU': 'X8NNQ9CYL2',
+        'UID': '6HQ443353U',
+    }
+    assert certificate.subject == expected_subject
+
+
+def test_issuer(certificate):
+    expected_issuer = {
+        'C': 'US',
+        'CN': 'Apple Worldwide Developer Relations Certification Authority',
+        'O': 'Apple Inc.',
+        'OU': 'Apple Worldwide Developer Relations',
+    }
+    assert certificate.issuer == expected_issuer
+
+
+def test_not_before(certificate):
+    assert certificate.not_before == '20191129113745Z'
+
+
+def test_not_after(certificate):
+    assert certificate.not_after == '20201128113745Z'
+
+
+def test_expires_at(certificate):
+    expected_expires_at = datetime(2020, 11, 28, 11, 37, 45, tzinfo=timezone.utc)
+    assert certificate.expires_at == expected_expires_at
+
+
+def test_has_expired(certificate):
+    assert certificate.has_expired is True
+
+
+def test_serial(certificate):
+    assert certificate.serial == 5308349992945343936
+
+
+def test_extensions(certificate):
+    expected_extensions = [
+        'Unknown OID',
+        'authorityInfoAccess',
+        'authorityKeyIdentifier',
+        'basicConstraints',
+        'certificatePolicies',
+        'extendedKeyUsage',
+        'keyUsage',
+        'subjectKeyIdentifier',
+    ]
+    assert sorted(certificate.extensions) == expected_extensions